external types
