+++ web-app/src/types.ts
@@ -1,0 +1,8 @@
+export type PhoneEntryData = {
+  phone_number: string;
+  extracted_name?: string;
+  url?: string;
+  survey_completed?: boolean;
+  out_of_target?: boolean;
+  saved_results?: boolean;
+}
